S2c freeze fix (#57)
* Three small bugfixes, one of which was causing real issues for MLab servers in developing nations.
A one-line compiler warning fix for test_c2s.
Fixing watchdog timer issues in web100srv.  (The server bug)
Making end-to-end unit tests less flaky by no longer just assuming that every port in the range [1024, 31024) is fine for the server to take as its own.

* Moved helper functions to testutils, so that they can be used by both c2s
and s2c. Used those new helper functions to fix a freezing error in s2c,
where the s2c test pauses until it hears back from the packet trace
engine, and so if the packet trace engine fails to start, then the s2c
test just freezes forever.

* Updated in response to comments by Josh to eliminate `strlen()`.
